fix(review): preserve lifecycle validation invariants
